update sample applicatio

+++ src/agnocastlib/src/agnocast_subscription.cpp
@@ -86,7 +86,7 @@ union ioctl_subscriber_args SubscriptionBase::initialize(bool is_take_sub)
   subscriber_args.qos_depth = (qos_.durability() == rclcpp::DurabilityPolicy::TransientLocal)
                                 ? static_cast<uint32_t>(qos_.depth())
                                 : 0;
-  subscriber_args.subscriber_pid = subscriber_pid;
+  subscriber_args.subscriber_pid = static_cast<uint32_t>(subscriber_pid);
   subscriber_args.init_timestamp = agnocast_get_timestamp();
   subscriber_args.is_take_sub = is_take_sub;
   if (ioctl(agnocast_fd, AGNOCAST_SUBSCRIBER_ADD_CMD, &subscriber_args) < 0) {
